+++ src/TestTrackingDiagrams.Extensions.SNS/SnsTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.SNS;
+
+public enum SnsTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}